+++ server/ingester/profile/decoder/decoder_parser.go
@@ -53,7 +53,7 @@ type processTracer struct {
 	pid       uint32
 	stime     int64
 	eventType string
-	value     uint32
+	value     uint64
 }
 
 // implement storage.Putter
